switch to FineWeb10B Edu dataset
I haven't run and checked if this is working fine. I am hoping it works.
Ill want to run it on a GPU but it will be very costly. Andrej has 8
A100 GPUs. He can fit 64 batches in his GPU and it took him approx 1.74
hours to run, 2 hrs lets say. For 1 GPU, 16 hrs :( !!